python-services/resume-parser/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymupdf==1.24.13
python-docx==1.1.2
openai==1.57.0
python-multipart==0.0.20
